make utils clients.py and migrate helxi and groq client to these -> more modular code
